Lnf moved form org.eclipse.riena.navigation.ui.swt to  org.eclipse.riena.ui.swt Some tests for status bar

+++ org.eclipse.riena.ui.ridgets.swt/src/org/eclipse/riena/internal/ui/ridgets/swt/StatusbarNumberRidget.java
@@ -49,7 +49,7 @@ public class StatusbarNumberRidget extends AbstractSWTRidget implements IStatusb
 			// TODO Numberformatter nutzen
 			numberString = number.toString();
 		}
-		getUIControl().setNumber(number.intValue());
+		getUIControl().setNumber(number);
 	}
 
 	/**
